+++ cli/test/commands/snapshot_command_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutterforge_ai_cli/flutterforge_cli.dart';
+import 'package:flutterforge_ai_cli/flutterforge_ai_cli.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';